Move ::BatchEditsHelper into the hyrax namespace
