+++ app/views/hyrax/collections/_list_works.html.erb
@@ -2,7 +2,7 @@
 
   <td>
     <label for="batch_document_<%= document.id %>" class="sr-only"><%=t("hyrax.dashboard.my.sr.batch_checkbox")%></label>
-    <%= render '/batch_select/add_button', document: document %>&nbsp;
+    <%= render 'hyrax/batch_select/add_button', document: document %>&nbsp;
   </td>
 
   <td>
@@ -29,7 +29,7 @@
     </div>
   </td>
 
-  <td><%= document.date_uploaded %></td>
+  <td class="date"><%= document.date_uploaded %></td>
   <td class='text-center'>
     <span class="fa <%= current_user.trophies.where(work_id: document.id).exists? ? 'fa-star highlighted-work' : 'fa-star-o trophy-off' %>"></span></td>
   <td><%= render_visibility_link document %></td>
