+++ app/views/hyrax/my/works/_list_works.html.erb
@@ -2,7 +2,7 @@
 
   <td>
     <label for="batch_document_<%= document.id %>" class="sr-only"><%=t("hyrax.dashboard.my.sr.batch_checkbox")%></label>
-    <%= render '/batch_select/add_button', document: document %>&nbsp;
+    <%= render 'hyrax/batch_select/add_button', document: document %>&nbsp;
   </td>
 
   <td>
