+++ spec/helpers/hyrax/batch_edits_helper_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe BatchEditsHelper, type: :helper do
+RSpec.describe Hyrax::BatchEditsHelper, type: :helper do
   describe "#render_check_all" do
     subject { helper.render_check_all }
     before do
